+++ merged-packages/eth-json-rpc-middleware/src/wallet.ts
@@ -156,14 +156,14 @@ WalletMiddlewareOptions): JsonRpcMiddleware<any, Block> {
 
   async function lookupAccounts(
     req: JsonRpcRequest,
-    res: PendingJsonRpcResponse<Json>,
+    res: PendingJsonRpcResponse,
   ): Promise<void> {
     res.result = await getAccounts(req);
   }
 
   async function lookupDefaultAccount(
     req: JsonRpcRequest,
-    res: PendingJsonRpcResponse<Json>,
+    res: PendingJsonRpcResponse,
   ): Promise<void> {
     const accounts = await getAccounts(req);
     res.result = accounts[0] || null;
@@ -175,7 +175,7 @@ WalletMiddlewareOptions): JsonRpcMiddleware<any, Block> {
 
   async function sendTransaction(
     req: JsonRpcRequest,
-    res: PendingJsonRpcResponse<Json>,
+    res: PendingJsonRpcResponse,
   ): Promise<void> {
     if (!processTransaction) {
       throw rpcErrors.methodNotSupported();
@@ -198,7 +198,7 @@ WalletMiddlewareOptions): JsonRpcMiddleware<any, Block> {
 
   async function signTransaction(
     req: JsonRpcRequest,
-    res: PendingJsonRpcResponse<Json>,
+    res: PendingJsonRpcResponse,
   ): Promise<void> {
     if (!processSignTransaction) {
       throw rpcErrors.methodNotSupported();
@@ -224,7 +224,7 @@ WalletMiddlewareOptions): JsonRpcMiddleware<any, Block> {
   //
   async function signTypedData(
     req: JsonRpcRequest,
-    res: PendingJsonRpcResponse<Json>,
+    res: PendingJsonRpcResponse,
   ): Promise<void> {
     if (!processTypedMessage) {
       throw rpcErrors.methodNotSupported();
@@ -259,7 +259,7 @@ WalletMiddlewareOptions): JsonRpcMiddleware<any, Block> {
 
   async function signTypedDataV3(
     req: JsonRpcRequest,
-    res: PendingJsonRpcResponse<Json>,
+    res: PendingJsonRpcResponse,
   ): Promise<void> {
     if (!processTypedMessageV3) {
       throw rpcErrors.methodNotSupported();
@@ -291,7 +291,7 @@ WalletMiddlewareOptions): JsonRpcMiddleware<any, Block> {
 
   async function signTypedDataV4(
     req: JsonRpcRequest,
-    res: PendingJsonRpcResponse<Json>,
+    res: PendingJsonRpcResponse,
   ): Promise<void> {
     if (!processTypedMessageV4) {
       throw rpcErrors.methodNotSupported();
@@ -323,7 +323,7 @@ WalletMiddlewareOptions): JsonRpcMiddleware<any, Block> {
 
   async function personalSign(
     req: JsonRpcRequest,
-    res: PendingJsonRpcResponse<Json>,
+    res: PendingJsonRpcResponse,
   ): Promise<void> {
     if (!processPersonalMessage) {
       throw rpcErrors.methodNotSupported();
@@ -380,7 +380,7 @@ WalletMiddlewareOptions): JsonRpcMiddleware<any, Block> {
 
   async function personalRecover(
     req: JsonRpcRequest,
-    res: PendingJsonRpcResponse<Json>,
+    res: PendingJsonRpcResponse,
   ): Promise<void> {
     if (
       !req?.params ||
@@ -403,7 +403,7 @@ WalletMiddlewareOptions): JsonRpcMiddleware<any, Block> {
 
   async function encryptionPublicKey(
     req: JsonRpcRequest,
-    res: PendingJsonRpcResponse<Json>,
+    res: PendingJsonRpcResponse,
   ): Promise<void> {
     if (!processEncryptionPublicKey) {
       throw rpcErrors.methodNotSupported();
@@ -425,7 +425,7 @@ WalletMiddlewareOptions): JsonRpcMiddleware<any, Block> {
 
   async function decryptMessage(
     req: JsonRpcRequest,
-    res: PendingJsonRpcResponse<Json>,
+    res: PendingJsonRpcResponse,
   ): Promise<void> {
     if (!processDecryptMessage) {
       throw rpcErrors.methodNotSupported();